// src/common/enums/demande-status.enum.ts
export enum DemandeStatus {
  BROUILLON = 'BROUILLON',
  SOUMISE = 'SOUMISE',
  INELIGIBLE = 'INELIGIBLE', // Demande acceptée mais l'agent ne remplit pas les prérequis
  EN_VALIDATION_HIERARCHIQUE = 'EN_VALIDATION_HIERARCHIQUE',
  VALIDEE_HIERARCHIQUE = 'VALIDEE_HIERARCHIQUE',
  REJETEE_HIERARCHIQUE = 'REJETEE_HIERARCHIQUE',
  EN_ETUDE_DGR = 'EN_ETUDE_DGR',
  AVIS_DGR_FAVORABLE = 'AVIS_DGR_FAVORABLE',
  AVIS_DGR_DEFAVORABLE = 'AVIS_DGR_DEFAVORABLE',
  EN_VERIFICATION_CVR = 'EN_VERIFICATION_CVR',
  VALIDEE_CVR = 'VALIDEE_CVR',
  REJETEE_CVR = 'REJETEE_CVR',
  EN_ETUDE_DNCF = 'EN_ETUDE_DNCF',
  ACCEPTEE = 'ACCEPTEE',
  REJETEE = 'REJETEE',
}